모든 코드 Authorization 적용 완료

+++ src/main/java/com/precapstone/fiveguys_backend/api/group/GroupService.java
@@ -1,5 +1,6 @@
 package com.precapstone.fiveguys_backend.api.group;
 
+import com.precapstone.fiveguys_backend.api.auth.JwtTokenProvider;
 import com.precapstone.fiveguys_backend.api.dto.groups.GroupsCreateDTO;
 import com.precapstone.fiveguys_backend.api.dto.groups.GroupsPatchDTO;
 import com.precapstone.fiveguys_backend.entity.Groups;
@@ -9,21 +10,24 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+import static com.precapstone.fiveguys_backend.exception.errorcode.BasicErrorCode.ACCESS_DENIED;
 import static com.precapstone.fiveguys_backend.exception.errorcode.GroupsErrorCode.*;
 
 @Service
 @RequiredArgsConstructor
 public class GroupService {
     private final GroupsRepository groupsRepository;
+    private final JwtTokenProvider jwtTokenProvider;
 
     /**
      * 그룹을 생성하는 함수
      * TODO: [예외처리] 그룹이 재귀되게 요청되는 경우
      *
      * @param groupsCreateDTO 그룹 생성 정보
+     * @param accessToken
      * @return
      */
-    public Groups createGroup(GroupsCreateDTO groupsCreateDTO) {
+    public Groups createGroup(GroupsCreateDTO groupsCreateDTO, String accessToken) {
         // 부모그룹ID를 통해 추가할 부모 그룹의 정보를 조회한다.
         var parentGroup = groupsRepository.findById(groupsCreateDTO.getParentGroupId())
                 .orElse(null);
@@ -34,6 +38,10 @@ public class GroupService {
          */
         if(parentGroup == null && groupsCreateDTO.getParentGroupId() != 0)
             throw new ControlledException(PARENT_GROUP_NOT_FOUND);
+
+        // [예외처리] 권한 소유자만 데이터를 반환 받을 수 있다.
+        if(!parentGroup.getUserId().equals(jwtTokenProvider.getUserIdFromToken(accessToken)))
+            throw new ControlledException(ACCESS_DENIED);
 
         /*
          * [예외처리] 올바르지 않은 그룹명 요청
@@ -46,6 +54,7 @@ public class GroupService {
 
         // 그룹 생성
         var group = Groups.builder()
+                .userId(accessToken)
                 .groupsName(groupsCreateDTO.getGroupsName())
                 .parent(parentGroup)
                 .build();
@@ -62,9 +71,14 @@ public class GroupService {
      * @param groupId 조회할 그룹ID
      * @return
      */
-    public Groups infoByGroupId(Long groupId) {
+    public Groups infoByGroupId(Long groupId, String accessToken) {
         var group = groupsRepository.findById(groupId)
                 .orElseThrow(() -> new ControlledException(GROUP_NOT_FOUND));
+
+        // [예외처리] 권한 소유자만 데이터를 반환 받을 수 있다.
+        if(!group.getUserId().equals(jwtTokenProvider.getUserIdFromToken(accessToken)))
+            throw new ControlledException(ACCESS_DENIED);
+
         return group;
     }
 
@@ -85,11 +99,16 @@ public class GroupService {
      * 그룹ID를 통해 본인의 하위 그룹을 조회한다.
      *
      * @param groupId
+     * @param accessToken
      * @return
      */
-    public List<Groups> childGroupInfo(Long groupId) {
+    public List<Groups> childGroupInfo(Long groupId, String accessToken) {
         var group = groupsRepository.findById(groupId)
                 .orElseThrow(() -> new ControlledException(GROUP_NOT_FOUND));
+
+        // [예외처리] 권한 소유자만 데이터를 반환 받을 수 있다.
+        if(!group.getUserId().equals(jwtTokenProvider.getUserIdFromToken(accessToken)))
+            throw new ControlledException(ACCESS_DENIED);
 
         var groups = groupsRepository.findByParent(group)
                 .orElseThrow(() -> new ControlledException(GROUP_NOT_FOUND));
@@ -116,21 +135,26 @@ public class GroupService {
     /**
      * 그룹을 삭제하는 함수
      * ※ 본인이 소유한 모든 하위 그룹들도 함께 사라진다.
-     * TODO: 반복문을 이용하지 않고 제거하는 방법을 고려해볼 것
+     * TODO: 재귀함수를 이용하지 않고 제거하는 방법을 고려해볼 것
      *
      * @param groupId 삭제할 그룹ID
      * @return
      */
-    public Groups deleteGroup(Long groupId) {
+    public Groups deleteGroup(Long groupId, String accessToken) {
         var group = groupsRepository
                 .findById(groupId)
                 .orElseThrow(() -> new ControlledException(GROUP_NOT_FOUND));
 
+        // [예외처리] 권한 소유자만 데이터를 반환 받을 수 있다.
+        if(!group.getUserId().equals(jwtTokenProvider.getUserIdFromToken(accessToken)))
+            throw new ControlledException(ACCESS_DENIED);
+
         // 본인이 소유한 자식 그룹 조회
         var childGroups = groupsRepository.findByParent(group)
                 .orElseThrow(() -> new ControlledException(GROUP_NOT_FOUND));
-        // 자식 그룹 모두 삭제
-        for (Groups child : childGroups) deleteGroup(child.getGroupsId());
+        
+        // 자식 그룹 모두 삭제(재귀함수)
+        for (Groups child : childGroups) deleteGroup(child.getGroupsId(), accessToken);
 
         // 해당 그룹 삭제
         groupsRepository.deleteById(groupId);
@@ -143,12 +167,17 @@ public class GroupService {
      * TODO: [예외처리] 그룹이 재귀되게 요청되는 경우
      *
      * @param groupsPatchDTO 그룹 변경 정보
+     * @param accessToken
      * @return
      */
-    public Groups updateGroup(GroupsPatchDTO groupsPatchDTO) {
+    public Groups updateGroup(GroupsPatchDTO groupsPatchDTO, String accessToken) {
         var group = groupsRepository
                 .findById(groupsPatchDTO.getGroupId())
                 .orElseThrow(() -> new ControlledException(GROUP_NOT_FOUND));
+
+        // [예외처리] 권한 소유자만 데이터를 반환 받을 수 있다.
+        if(!group.getUserId().equals(jwtTokenProvider.getUserIdFromToken(accessToken)))
+            throw new ControlledException(ACCESS_DENIED);
 
         // 특정 인자가 null로 반환된 경우 수정정보가 저장되지 않는다.
         if(groupsPatchDTO.getNewGroupName() != null) group.setGroupsName(groupsPatchDTO.getNewGroupName());
